requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
Flask-Session==0.5.0
Jinja2==3.1.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pipenv==2024.0.1
python-dotenv==1.0.1
requests==2.31.0
